train_test

+++ models/sam_DA.py
@@ -588,7 +588,7 @@ class SAM(nn.Module):
         #2
         else:
             dis_gt = self.netD_mask(self.gt_mask)
-            dis_offset = self.netD_offset(offset_gt)
+            dis_offset = self.netD_offset(offset_gt.float())
             space_query1 = self.netD_mask(self.pred_mask.detach())
             space_query2 = self.netD_mask(self.pred_mask2.detach())
             channel_query1 = self.netD_offset(self.masks_hq.detach())
